refactor: dynamically import code editor modal
